% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praise_me}
\alias{praise_me}
\alias{praise_someone}
\title{Praise someone!}
\usage{
praise_me()

praise_someone(who = NULL)
}
\arguments{
\item{who}{A character of who to praise.}
}
\description{
This function praises a person.
}
\examples{
praise_me()
praise_someone("Stella")

}
